Combined alike get/setters in files
