+++ backend/src/handleAlignments.js
@@ -1,8 +1,6 @@
 import { readFile } from "fs/promises";
 
-const getAlignments = async () => {
+export const getAlignments = async () => {
   const alignments = JSON.parse(await readFile("data/alignments.json"));
   return alignments;
 };
-
-export default getAlignments;
